+++ BlizuTebe/Dtos/UserDto.cs
@@ -2,7 +2,7 @@
 {
     public class UserDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Username { get; set; }
         public string Password { get; set; }
         public string Name { get; set; }
@@ -10,7 +10,7 @@
         public UserRole Role { get; set; }
         public DateTime DateOfBirth { get; set; }
         public Boolean IsVerified { get; set; }
-        public int? LocalCommunityId { get; set; }
+        public long? LocalCommunityId { get; set; }
         public IFormFile? Picture { get; set; }
         public string? ProfilePicture { get; set; }
         public double Rating { get; set; }
