+++ BlizuTebe/Models/Announcement.cs
@@ -2,19 +2,19 @@
 {
     public class Announcement
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
         public string Picture { get; set; }
         public DateTime PublishedAt { get; set; }
         public DateTime ExpirationDate { get; set; }
         public bool IsImportant { get; set; }
-        public int? AdminId { get; set; }
-        public int? LocalCommuntyId { get; set; }
+        public long? AdminId { get; set; }
+        public long? LocalCommuntyId { get; set; }
 
         public Announcement() { }
 
-        public Announcement(int id, string title, string description, string picture, DateTime publishedAt, DateTime expirationDate, bool isImportant, int? adminId, int? localCommuntyId)
+        public Announcement(long id, string title, string description, string picture, DateTime publishedAt, DateTime expirationDate, bool isImportant, long? adminId, long? localCommuntyId)
         {
             Id = id;
             Title = title;
