+++ BlizuTebe/Models/LocalCommunity.cs
@@ -5,18 +5,18 @@ namespace BlizuTebe.Models
 {
     public class LocalCommunity
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string City { get; set; }
         public Geometry Boundary { get; set; }
         public NetTopologySuite.Geometries.Point CenterPoint { get; set; }
-        public int? PresidentId { get; set; }
+        public long? PresidentId { get; set; }
         public string? PhoneNumber { get; set; }
         public string? Facebook { get; set; }
 
         public LocalCommunity() { }
 
-        public LocalCommunity(int id, string name, string city, Geometry boundary, NetTopologySuite.Geometries.Point centerPoint, int? presidentId, string phoneNumber, string facebook)
+        public LocalCommunity(long id, string name, string city, Geometry boundary, NetTopologySuite.Geometries.Point centerPoint, long? presidentId, string phoneNumber, string facebook)
         {
             Id = id;
             Name = name;
