+++ BlizuTebe/Models/User.cs
@@ -2,7 +2,7 @@
 {
     public class User
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Username { get; set; }
         public string Password { get; set; }
         public string Name { get; set; }
@@ -10,13 +10,13 @@
         public UserRole Role { get; set; }
         public DateTime DateOfBirth { get; set; }
         public Boolean IsVerified { get; set; }
-        public int? LocalCommunityId { get; set; }
+        public long? LocalCommunityId { get; set; }
         public string? ProfilePicture { get; set; }
         public double Rating { get; set; }
 
         public User() { }
 
-        public User(int id, string username, string password, string name, string surname, UserRole role, DateTime dateOfBirth, bool isVerified, int? localCommunityId, string profilePicture, double rating)
+        public User(long id, string username, string password, string name, string surname, UserRole role, DateTime dateOfBirth, bool isVerified, long? localCommunityId, string profilePicture, double rating)
         {
             Id = id;
             Username = username;
